+++ fe1-web/components/RecorderButton.tsx
@@ -5,7 +5,7 @@ import {
 import PropTypes from 'prop-types';
 
 import { Colors } from 'styles';
-import circularButtonStyles from 'styles/stylesheets/circlarButtonStyles';
+import circularButtonStyles from '../styles/stylesheets/circularButtonStyles';
 
 /**
  * Recorder button that executes an onPress action given in props
