+++ fe1-web/model/network/validation/dataSchemas.ts
