+++ fe1-web/model/network/validation/querySchemas.ts
